#20 Firestoreに読み書きする処理はDataServiceに切り分けた(AuthはエラーハンドリングがあるためNotifierに残した)
