+++ lib/data_service/person_data_service.dart
@@ -1,0 +1,30 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:ohayo_post_app/model/person.dart';
+
+class PersonDataService {
+  final FirebaseFirestore _storeDB = FirebaseFirestore.instance;
+
+  Future<void> createPerson(String uid, String nickName, String email) async {
+    final _personMap = {
+      'uid': uid,
+      'nickName': nickName,
+      'email': email,
+      'createAt': Timestamp.now(),
+    };
+    _storeDB.collection('persons').doc(uid).set(_personMap);
+  }
+
+  Stream<Person> personStream(String uid) {
+    return _storeDB
+        .collection('persons')
+        .doc(uid)
+        .snapshots()
+        .map((DocumentSnapshot data) => Person.fromDocumentSnapshot(data));
+  }
+
+  Future<Person> getPerson(String uid) async {
+    if (uid.isEmpty) return Person.empty();
+    final doc = await _storeDB.collection('persons').doc(uid).get();
+    return Person.fromDocumentSnapshot(doc);
+  }
+}
